上架設備 --- OK
功能待進一步優化

+++ UnityProject/iDCIM_Transglobal/Assets/_VictorDev/TCIT/Revit/RevitModelData.cs
@@ -300,8 +300,8 @@ namespace _VictorDEV.Revit
         public float width;
         public float height;
         public int heightU;
-        public float watt;
-        public float weight;
+        public int watt;
+        public int weight;
     }
 
     #endregion
